refactor(App): App 을 ConsoleApp 으로 변경하여 콘솔 로또와 웹 로또를 분리
